Fetch Pending Commits (#533)
* Combine tracerAPI calls into one folder

* Remove unused hook

* Update foreach loop

* Collapse unescessary wrapper

* Allow undefined networks

* fix trade history endpoint
