+++ libs/utils/supportedNetworks.ts
@@ -1,6 +1,6 @@
 import { KnownNetwork, NETWORKS } from '@tracer-protocol/pools-js';
 
-export const isSupportedNetwork = (networkId?: KnownNetwork): boolean => {
+export const isSupportedNetwork = (networkId: number | string | undefined): boolean => {
     const networkIdString = networkId?.toString();
 
     return networkIdString === NETWORKS.ARBITRUM || networkIdString === NETWORKS.ARBITRUM_RINKEBY;
